+++ src/app/core/services/package.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 import {
-  PackageList,
   PackageVersion,
   Package,
   deserializablePackageList
@@ -12,6 +11,12 @@ import { PreferencesService } from './preferences.service';
 import { ReadStream } from 'fs';
 import { ThunderstoreService } from './thunderstore.service';
 import { DatabaseService } from './database.service';
+import { SelectionModel } from '@angular/cdk/collections';
+import { Selectable } from '../models/selectable.model';
+
+export interface SelectablePackge extends Selectable, Package {}
+
+import { protocols } from '../../../../package.json';
 
 export interface PackageChangeset {
   updated: Set<PackageVersion>;
@@ -27,11 +32,14 @@ const BEPIN_UUID4 = '4c253b36-fd0b-4e6d-b4d8-b227972af4da';
 
 @Injectable()
 export class PackageService {
-  private installedPackagesSource = new BehaviorSubject<PackageList>([]);
+  private installedPackagesSource = new BehaviorSubject<SelectablePackge[]>([]);
   public installedPackages$ = this.installedPackagesSource.asObservable();
 
-  private allPackagesSource = new BehaviorSubject<PackageList>([]);
+  private allPackagesSource = new BehaviorSubject<SelectablePackge[]>([]);
   public allPackages$ = this.allPackagesSource.asObservable();
+
+  public selectedPackage = new BehaviorSubject<Package>(undefined);
+  public selection = new SelectionModel<SelectablePackge>(true, []);
 
   constructor(
     private download: DownloadService,
@@ -40,6 +48,8 @@ export class PackageService {
     private thunderstore: ThunderstoreService,
     private db: DatabaseService
   ) {
+    this.registerHttpProtocol();
+
     if (this.prefs.get('checkUpdatesOnStart')) {
       this.downloadPackageList();
     } else {
@@ -58,7 +68,7 @@ export class PackageService {
     }
   }
 
-  public async loadPackagesFromCache(): Promise<PackageList> {
+  public async loadPackagesFromCache(): Promise<SelectablePackge[]> {
     const serializedPackages = await this.db.packageTable.toArray();
 
     const packages = deserializablePackageList(serializedPackages);
@@ -72,7 +82,7 @@ export class PackageService {
     return packages;
   }
 
-  public downloadPackageList(): Observable<PackageList> {
+  public downloadPackageList(): Observable<SelectablePackge[]> {
     const oldPackages = this.allPackagesSource.value;
     this.allPackagesSource.next(null);
 
@@ -266,5 +276,39 @@ export class PackageService {
       'BepInEx',
       'plugins'
     );
+  }
+
+  private registerHttpProtocol() {
+    for (const scheme of protocols) {
+      console.log(`Registering protocol ${scheme}`);
+      this.electron.protocol.registerHttpProtocol(scheme, (req, cb) => {
+        // format ror2mm://v1/install/thunderstore.io/[author]/[package]/[version]/
+        const chunks = req.url.split('/');
+        // const protocol = chunks[0];
+        const [
+          protocolVersion,
+          action,
+          provider,
+          author,
+          packageName,
+          version
+        ] = chunks.slice(2);
+        if (
+          protocolVersion === 'v1' &&
+          action === 'install' &&
+          provider === 'thunderstore.io'
+        ) {
+          const packageToInstall = this.allPackagesSource.value.find(
+            p => p.owner === author && p.name === packageName
+          );
+          const versionToInstall = packageToInstall.versions.find(
+            v => v.version.version === version
+          );
+          console.log('Marking package for install', versionToInstall);
+          this.selection.select(versionToInstall.pkg);
+          this.selectedPackage.next(packageToInstall);
+        }
+      });
+    }
   }
 }