+++ src/app/nav-menu/nav-menu/nav-menu.component.ts
@@ -15,7 +15,7 @@ import { version } from '../../../../package.json';
 })
 export class NavMenuComponent implements OnInit {
   isHandset$: Observable<boolean> = this.breakpointObserver
-    .observe(Breakpoints.Small)
+    .observe([Breakpoints.Small, Breakpoints.XSmall])
     .pipe(map(result => result.matches));
 
   version: string;