+++ src/app/package-selection/packages-page/packages-page.component.ts
@@ -4,7 +4,7 @@ import {
   PackageChangeset
 } from '../../core/services/package.service';
 import { Observable } from 'rxjs';
-import { Package, PackageList } from '../../core/models/package.model';
+import { PackageList } from '../../core/models/package.model';
 import { map, tap } from 'rxjs/operators';
 
 @Component({
@@ -16,7 +16,6 @@ import { map, tap } from 'rxjs/operators';
   }
 })
 export class PackagesPageComponent {
-  selectedPackage: Package;
   installedPackages$: Observable<
     PackageList
   > = this.service.installedPackages$.pipe(